rename canvas dashboard wrapper
